Выполнены задачи согласно JB-46 (UPD2): - Выпилил прилетевший код работы с избранными вакансиями, оставил свой в рамках этой задачи - Исправил двойную строчку с kapt в gradle - Скорректированы зависимости проекта для корректной работы библиотеки Room - отсутствовала зависимость для kapt - Обновлена DAO модель для хранения вакансии в БД - Созданы интерфейсы и реализованы методы для репозитория и интерактора для работы с базой данных избранных вакансий - Реализован конвертер данных между моделью представления и моделью хранения вакансии в БД
