+++ app/src/main/java/ru/practicum/android/diploma/data/db/VacancyFavoriteEntity.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.data.dto
+package ru.practicum.android.diploma.data.db
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
