+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -15,7 +15,7 @@ val interactorModule = module {
     }
 
     single<VacanciesInteractor> {
-        VacanciesInteractorImpl(get(), get(), get())
+        VacanciesInteractorImpl(get(), get())
     }
 
     single<FavoritesInteractor> {
